default age for guards making them not show after 36 hours
If never set by user, the app displays to the user that the age is unlimited, but in reality was only set for 36 hours. Changing the default to 900 hours if no prefkey is set which should cover well over the time period for swgcraft to automatically despawn an old resource.