hiding asteroids from test bench
since they are always the same in game, and never shown as an available resource on craft, they can not be ever used in a test bench, so I'm removing them, thusly treated as a sub component. Should not hurt the math at all since it's an elimination of a capped static resource.