removing image album
While it's a neat novelty it seems superfluous. The folks I've asked don't seem to use it. I don't see the point in duplicating screenshot files as it's a waste of resources, and modern systems make it very easy to view them in the SWG screenshot folder anyway. So for both performance and code maintaining reasons I'm removing this entirely.