fix for inventory issues
Remodeled the inventory map. For whatever reason SWGCGalaxy object is not a good thing to key on. It was missing hits on get(key) lookups causing duplicates to be created.
Changed model to just use String and SWGCGalaxy.toString()
With the implemention of equals overrides in inventory wrappers, this should completely fix issues with inventory and refreshing harvesters.

This will require an upgrade routine be written due to the change in the underlying hashmap that is stored in the DAT file.