more logic for hopper and power
If power runs out, hopper stops filling.
If resource is depleted, hopper stops filling on depletion time. Note that depletion times in aide are best guess only. If aide has not been ran for a few days, and a resource was depleted during that time, aide only sees it as depleted the next time it pulls an update.
If hopper is full power stops counting down.